feat: adds null migrator
