+++ bottlerocket-settings-sdk/src/migrate/null/extensionbuilder.rs
@@ -1,0 +1,9 @@
+use super::NullMigrator;
+use crate::extension_builder;
+
+extension_builder!(
+    pub,
+    NullMigratorExtensionBuilder,
+    NullMigrator,
+    NullMigrator
+);
